+++ SSLTest/src/scan_parameters/non_ratable/port_discovery.py
@@ -4,8 +4,6 @@ import nmap3
 import requests
 import urllib3
 
-from ...utils import incremental_sleep
-
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 log = logging.getLogger(__name__)
 
@@ -18,29 +16,25 @@ def discover_ports(url):
     :return: Usable ports
     :rtype: list
     """
-    log.info('Discovering ports')
-    nmap = nmap3.NmapHostDiscovery()
+    log.info('Discovering ports with nmap')
+    nmap = nmap3.NmapScanTechniques()
     # Scan with nmap for all open ports
-    result = nmap.nmap_portscan_only(url)
+    result = nmap.nmap_syn_scan(url, args="-p1-65535")
     open_ports = [port['portid'] for port in list(result.items())[0][1]['ports']]
     usable_ports = []
-    log.debug(f'Scanned ports by nmap : {open_ports}')
+    log.info(f'Scanned ports by nmap : {open_ports}')
     for port in open_ports:
-        sleep = 0
-        # Loop until there is a valid response or after 10 seconds
         while True:
             try:
-                head = requests.head(f'https://{url}:{port}', timeout=5, verify=False,
+                head = requests.head(f'https://{url}:{port}', timeout=3, verify=False,
                                      headers={'Connection': 'close'})
                 if head.status_code < 400:
                     usable_ports.append(int(port))
+                    log.debug(f"Port {port} valid")
                     break
-            except requests.exceptions.SSLError:
+            except (requests.exceptions.ReadTimeout, requests.exceptions.Timeout, requests.exceptions.ConnectionError,
+                    requests.exceptions.SSLError):
+                log.debug(f"Port {port} invalid")
                 break
-            # Valid exception, there might a port alive after timeout
-            except (requests.exceptions.ReadTimeout, requests.exceptions.Timeout):
-                usable_ports.append(int(port))
-                break
-            except requests.exceptions.ConnectionError:
-                break
+    log.info(f"Found ports: {usable_ports}")
     return usable_ports