+++ SSLTest/src/scan_vulnerabilities/tests/Crime.py
@@ -6,7 +6,8 @@ from ...utils import send_data_return_sock, is_server_hello
 
 
 class Crime(VulnerabilityTest):
-    test_name = 'CRIME'
+    name = short_name = 'CRIME'
+    description = "Test for ssl/tls encoding methods"
 
     def __init__(self, supported_protocols, address, timeout, protocol):
         super().__init__(supported_protocols, address, timeout, protocol)
@@ -26,7 +27,7 @@ class Crime(VulnerabilityTest):
             0x01  # Compression method (deflate)
         ])
         client_hello = client_hello.construct_client_hello()
-        response, sock = send_data_return_sock(self.address, client_hello, self.timeout, self.test_name)
+        response, sock = send_data_return_sock(self.address, client_hello, self.timeout, self.name)
         sock.close()
         if not is_server_hello(response):
             return False